feat: move the logic of the web app script to a command line script
